no need for special index names update timestamps to bigInteger

+++ app/modules/displays_info/migrations/2017_02_09_234251_displays.php
@@ -18,7 +18,7 @@ class Displays extends Migration
             $table->string('model')->nullable();
             $table->string('manufactured')->nullable();
             $table->string('native')->nullable();
-            $table->integer('timestamp')->nullable();
+            $table->bigInteger('timestamp')->nullable();
 
 //            $table->timestamps();
         });
